init Document Validation
