+++ src/Data/Morpheus/Client.hs
@@ -17,11 +17,11 @@ import qualified Data.ByteString.Lazy                    as L (readFile)
 import           Language.Haskell.TH
 import           Language.Haskell.TH.Quote
 
+import           Data.Morpheus.Document                  (parseFullGQLDocument)
 -- MORPHEUS
 import           Data.Morpheus.Execution.Client.Build    (defineQuery)
 import           Data.Morpheus.Execution.Client.Compile  (compileSyntax)
 import           Data.Morpheus.Execution.Client.Fetch    (Fetch (..))
-import           Data.Morpheus.Parsing.Document.Parse    (parseFullGQLDocument)
 import           Data.Morpheus.Parsing.JSONSchema.Parse  (decodeIntrospection)
 import           Data.Morpheus.Types.Internal.Data       (DataTypeLib)
 import           Data.Morpheus.Types.Internal.Validation (Validation)
