+++ src/Data/Morpheus/Validation/Query/Fragment.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE NamedFieldPuns #-}
 
-module Data.Morpheus.Validation.Fragment
+module Data.Morpheus.Validation.Query.Fragment
   ( validateFragments
   , castFragmentType
   , resolveSpread
@@ -20,15 +20,12 @@ import           Data.Morpheus.Types.Internal.AST.RawSelection (Fragment (..), F
 import           Data.Morpheus.Types.Internal.Base             (EnhancedKey (..), Position)
 import           Data.Morpheus.Types.Internal.Data             (DataTypeLib)
 import           Data.Morpheus.Types.Internal.Validation       (Validation)
-import           Data.Morpheus.Validation.Utils.Utils          (checkNameCollision, existsObjectType)
+import           Data.Morpheus.Validation.Internal.Utils       (checkNameCollision, existsObjectType)
 
-validateFragments ::
-     DataTypeLib -> FragmentLib -> [(Text, RawSelection)] -> Validation ()
-validateFragments lib fragments operatorSel =
-  validateNameCollision >> checkLoop >> checkUnusedFragments
+validateFragments :: DataTypeLib -> FragmentLib -> [(Text, RawSelection)] -> Validation ()
+validateFragments lib fragments operatorSel = validateNameCollision >> checkLoop >> checkUnusedFragments
   where
-    validateNameCollision =
-      checkNameCollision fragmentsKeys fragmentNameCollision
+    validateNameCollision = checkNameCollision fragmentsKeys fragmentNameCollision
     checkUnusedFragments =
       case fragmentsKeys \\ usedFragments fragments operatorSel of
         []     -> return ()
@@ -36,8 +33,7 @@ validateFragments lib fragments operatorSel =
     checkLoop = mapM (validateFragment lib) fragments >>= detectLoopOnFragments
     fragmentsKeys = map toEnhancedKey fragments
       where
-        toEnhancedKey (key, Fragment {fragmentPosition}) =
-          EnhancedKey key fragmentPosition
+        toEnhancedKey (key, Fragment {fragmentPosition}) = EnhancedKey key fragmentPosition
 
 type Node = EnhancedKey
 
@@ -51,62 +47,41 @@ getFragment Reference {referenceName, referencePosition} lib =
     Nothing       -> Left $ unknownFragment referenceName referencePosition
     Just fragment -> pure fragment
 
-castFragmentType ::
-     Maybe Text -> Position -> [Text] -> Fragment -> Validation Fragment
+castFragmentType :: Maybe Text -> Position -> [Text] -> Fragment -> Validation Fragment
 castFragmentType key' position' targets' fragment@Fragment {fragmentType} =
   if fragmentType `elem` targets'
     then pure fragment
-    else Left $
-         cannotBeSpreadOnType key' fragmentType position' (T.concat targets')
+    else Left $ cannotBeSpreadOnType key' fragmentType position' (T.concat targets')
 
 resolveSpread :: FragmentLib -> [Text] -> Reference -> Validation Fragment
-resolveSpread fragments allowedTargets reference@Reference { referenceName
-                                                           , referencePosition
-                                                           } =
-  getFragment reference fragments >>=
-  castFragmentType (Just referenceName) referencePosition allowedTargets
+resolveSpread fragments allowedTargets reference@Reference {referenceName, referencePosition} =
+  getFragment reference fragments >>= castFragmentType (Just referenceName) referencePosition allowedTargets
 
 usedFragments :: FragmentLib -> [(Text, RawSelection)] -> [Node]
 usedFragments fragments = concatMap findAllUses
   where
     findAllUses :: (Text, RawSelection) -> [Node]
-    findAllUses (_, RawSelectionSet RawSelection' {rawSelectionRec}) =
-      concatMap findAllUses rawSelectionRec
-    findAllUses (_, RawAlias {rawAliasSelection}) =
-      concatMap findAllUses [rawAliasSelection]
-    findAllUses (_, InlineFragment Fragment {fragmentSelection}) =
-      concatMap findAllUses fragmentSelection
+    findAllUses (_, RawSelectionSet RawSelection' {rawSelectionRec}) = concatMap findAllUses rawSelectionRec
+    findAllUses (_, RawAlias {rawAliasSelection}) = concatMap findAllUses [rawAliasSelection]
+    findAllUses (_, InlineFragment Fragment {fragmentSelection}) = concatMap findAllUses fragmentSelection
     findAllUses (_, RawSelectionField {}) = []
     findAllUses (_, Spread Reference {referenceName, referencePosition}) =
       [EnhancedKey referenceName referencePosition] <> searchInFragment
       where
-        searchInFragment =
-          maybe
-            []
-            (concatMap findAllUses . fragmentSelection)
-            (lookup referenceName fragments)
+        searchInFragment = maybe [] (concatMap findAllUses . fragmentSelection) (lookup referenceName fragments)
 
 scanForSpread :: (Text, RawSelection) -> [Node]
-scanForSpread (_, RawSelectionSet RawSelection' {rawSelectionRec = selection'}) =
-  concatMap scanForSpread selection'
-scanForSpread (_, RawAlias {rawAliasSelection = selection'}) =
-  concatMap scanForSpread [selection']
-scanForSpread (_, InlineFragment Fragment {fragmentSelection = selection'}) =
-  concatMap scanForSpread selection'
+scanForSpread (_, RawSelectionSet RawSelection' {rawSelectionRec = selection'}) = concatMap scanForSpread selection'
+scanForSpread (_, RawAlias {rawAliasSelection = selection'}) = concatMap scanForSpread [selection']
+scanForSpread (_, InlineFragment Fragment {fragmentSelection = selection'}) = concatMap scanForSpread selection'
 scanForSpread (_, RawSelectionField {}) = []
-scanForSpread (_, Spread Reference { referenceName = name'
-                                   , referencePosition = position'
-                                   }) = [EnhancedKey name' position']
+scanForSpread (_, Spread Reference {referenceName = name', referencePosition = position'}) =
+  [EnhancedKey name' position']
 
 validateFragment :: DataTypeLib -> (Text, Fragment) -> Validation NodeEdges
-validateFragment lib (fName, Fragment { fragmentSelection
-                                      , fragmentType
-                                      , fragmentPosition
-                                      }) =
+validateFragment lib (fName, Fragment {fragmentSelection, fragmentType, fragmentPosition}) =
   existsObjectType fragmentPosition fragmentType lib >>
-  pure
-    ( EnhancedKey fName fragmentPosition
-    , concatMap scanForSpread fragmentSelection)
+  pure (EnhancedKey fName fragmentPosition, concatMap scanForSpread fragmentSelection)
 
 detectLoopOnFragments :: Graph -> Validation ()
 detectLoopOnFragments lib = mapM_ checkFragment lib
