+++ src/Data/Morpheus/Validation/Query/Validation.hs
@@ -3,7 +3,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeOperators       #-}
 
-module Data.Morpheus.Validation.Validation
+module Data.Morpheus.Validation.Query.Validation
   ( validateRequest
   ) where
 
@@ -15,15 +15,13 @@ import           Data.Morpheus.Types.Internal.AST.Operation (Operation (..), Ope
 import           Data.Morpheus.Types.Internal.Data          (DataOutputObject, DataTypeLib (..))
 import           Data.Morpheus.Types.Internal.Validation    (Validation)
 import           Data.Morpheus.Types.Types                  (GQLQueryRoot (..))
-import           Data.Morpheus.Validation.Fragment          (validateFragments)
-import           Data.Morpheus.Validation.Selection         (validateSelectionSet)
-import           Data.Morpheus.Validation.Utils.Utils       (VALIDATION_MODE)
-import           Data.Morpheus.Validation.Variable          (resolveOperationVariables)
+import           Data.Morpheus.Validation.Internal.Utils    (VALIDATION_MODE)
+import           Data.Morpheus.Validation.Query.Fragment    (validateFragments)
+import           Data.Morpheus.Validation.Query.Selection   (validateSelectionSet)
+import           Data.Morpheus.Validation.Query.Variable    (resolveOperationVariables)
 
-getOperationDataType ::
-     RawOperation -> DataTypeLib -> Validation DataOutputObject
-getOperationDataType Operation {operationKind = QUERY} lib =
-  pure $ snd $ query lib
+getOperationDataType :: RawOperation -> DataTypeLib -> Validation DataOutputObject
+getOperationDataType Operation {operationKind = QUERY} lib = pure $ snd $ query lib
 getOperationDataType Operation {operationKind = MUTATION, operationPosition} lib =
   case mutation lib of
     Just (_, mutation') -> pure mutation'
@@ -33,8 +31,7 @@ getOperationDataType Operation {operationKind = SUBSCRIPTION, operationPosition}
     Just (_, subscription') -> pure subscription'
     Nothing                 -> Left $ subscriptionIsNotDefined operationPosition
 
-validateRequest ::
-     DataTypeLib -> VALIDATION_MODE -> GQLQueryRoot -> Validation ValidOperation
+validateRequest :: DataTypeLib -> VALIDATION_MODE -> GQLQueryRoot -> Validation ValidOperation
 validateRequest lib validationMode GQLQueryRoot { fragments
                                                 , inputVariables
                                                 , operation = rawOperation@Operation { operationName
@@ -44,27 +41,7 @@ validateRequest lib validationMode GQLQueryRoot { fragments
                                                                                      }
                                                 } = do
   operationDataType <- getOperationDataType rawOperation lib
-  variables <-
-    resolveOperationVariables
-      lib
-      fragments
-      (fromList inputVariables)
-      validationMode
-      rawOperation
+  variables <- resolveOperationVariables lib fragments (fromList inputVariables) validationMode rawOperation
   validateFragments lib fragments operationSelection
-  selection <-
-    validateSelectionSet
-      lib
-      fragments
-      operationName
-      variables
-      operationDataType
-      operationSelection
-  pure $
-    Operation
-      { operationName
-      , operationKind
-      , operationArgs = []
-      , operationSelection = selection
-      , operationPosition
-      }
+  selection <- validateSelectionSet lib fragments operationName variables operationDataType operationSelection
+  pure $ Operation {operationName, operationKind, operationArgs = [], operationSelection = selection, operationPosition}
